requirements.txt
flask==2.0.1
werkzeug==2.0.1
torch==1.9.0
torchvision==0.10.0
pillow==8.3.1
numpy==1.21.1
pytest==6.2.5
